+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Add `:round_robin` dispatch strategy
+
 ## [0.3.1] - 2026-05-12
 
 - Fix `result_port` race condition in single-worker shutdown